move consistent_hashing to hashing
